Add files via upload
This triple data-set is an expansion of my idea from last mod.  Each of the discs have been updated to include photos and plastic types.  The 2nd dataset is a description of each type of disc category, and the plastic types that they are frequently used in.  The final dataset is a list of each type of plastic and their benefits.
These could be used to create an interface for cross referencing which discs are available in which plastics, other discs of the same sort to be referenced.